+++ src/Datasets/Generators/HalfMoon.php
@@ -62,17 +62,17 @@ class HalfMoon implements Generator
         float $rotation = 90.0,
         float $noise = 0.1
     ) {
-        if ($scale < 0.) {
+        if ($scale < 0.0) {
             throw new InvalidArgumentException('Scaling factor must be greater'
                 . " than 0, $scale given.");
         }
 
-        if ($rotation < 0. or $rotation > 360.) {
+        if ($rotation < 0.0 or $rotation > 360.0) {
             throw new InvalidArgumentException('Rotation must be between 0 and'
                 . " 360 degrees, $rotation given.");
         }
 
-        if ($noise < 0.) {
+        if ($noise < 0.0) {
             throw new InvalidArgumentException('Noise factor must be greater'
                 . " than 0, $noise given.");
         }